requirements.txt
ruamel.yaml
mako

sphinx
sphinx-autodoc-typehints
sphinx-autoapi